+++ packages/form-control/src/components/base-form-control/Component.tsx
@@ -196,8 +196,6 @@ export const BaseFormControl = React.forwardRef<HTMLDivElement, BaseFormControlP
                     styles[`size-${size}`],
                     {
                         [commonStyles.block]: block,
-                        [commonStyles.hasLeftAddons]: leftAddons,
-                        [commonStyles.hasRightAddons]: rightAddons || error,
                     },
                 )}
             >