+++ packages/input/src/shared/index.ts
@@ -1,4 +1,3 @@
 export * from '../components/clear-button';
 export { getInputTestIds } from '../utils';
 export { LockIcon } from '../components/lock-icon';
-export { getAddonsByPriority } from '../helpers/get-addons-by-priority';