+++ packages/number-input/src/components/number-input/Component.tsx
@@ -3,6 +3,7 @@ import React, {
     type FocusEvent,
     forwardRef,
     type ForwardRefExoticComponent,
+    Fragment,
     type RefAttributes,
     useEffect,
     useMemo,
@@ -13,7 +14,6 @@ import { type MaskitoOptions, maskitoTransform } from '@maskito/core';
 import { useMaskito } from '@maskito/react';
 
 import { type InputProps } from '@alfalab/core-components-input';
-import { getAddonsByPriority } from '@alfalab/core-components-input/shared';
 import { fnUtils, isIOS } from '@alfalab/core-components-shared';
 
 import {
@@ -227,16 +227,10 @@ export const NumberInput = forwardRef<HTMLInputElement, NumberInputProps>(
             setFocused(false);
         };
 
-        const rightAddonsMap = getAddonsByPriority([
-            {
-                priority: 2,
-                predicate: Boolean(rightAddons),
-                render: () => rightAddons,
-            },
-            {
-                priority: 1,
-                predicate: withStepper && !disabled && !readOnly,
-                render: () => (
+        const renderRightAddons = () => (
+            <Fragment>
+                {rightAddons}
+                {withStepper && !disabled && !readOnly && (
                     <Steppers
                         colors={colors}
                         dataTestId={dataTestId}
@@ -249,9 +243,9 @@ export const NumberInput = forwardRef<HTMLInputElement, NumberInputProps>(
                         onDecrement={handleDecrement}
                         size={size}
                     />
-                ),
-            },
-        ]);
+                )}
+            </Fragment>
+        );
 
         return (
             <Input
@@ -270,7 +264,7 @@ export const NumberInput = forwardRef<HTMLInputElement, NumberInputProps>(
                 size={size}
                 disableUserInput={disableUserInput}
                 clear={clearProp && /\d/.test(value)}
-                rightAddons={rightAddonsMap}
+                rightAddons={renderRightAddons()}
             />
         );
     },