+++ packages/button/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @alfalab/core-components-button
 
+## 13.1.1
+
+### Patch Changes
+
+<sup><time>06.03.2026</time></sup>
+
+#### [#2102](https://github.com/core-ds/core-components/pull/2102)
+
+##### Vars
+
+- Удален файл `no-typography-index.css`
+
+- Обновлены зависимости
+    - @alfalab/core-components-spinner@6.0.3
+    - @alfalab/core-components-shared@2.1.1
+    - @alfalab/core-components-mq@6.0.3
+
+## 13.1.0
+
+### Minor Changes
+
+<sup><time>27.02.2026</time></sup>
+
+#### [#2086](https://github.com/core-ds/core-components/pull/2086)
+
+##### Button
+
+- Добавлены `css` переменные для темизации подписи под лейблом (хинта):
+    - `--button-accent-hint-color`
+    - `--button-inverted-accent-hint-color`
+    - `--button-mobile-primary-hint-color`
+    - `--button-mobile-inverted-primary-hint-color`
+    - `--button-mobile-accent-hint-color`
+    - `--button-mobile-inverted-accent-hint-color`
+
+### Patch Changes
+
+<sup><time>27.02.2026</time></sup>
+
+#### [#2086](https://github.com/core-ds/core-components/pull/2086)
+
+##### Button
+
+- Исправлено отображение подписи под лейблом (хинта) в заблокированном состоянии
+
 ## 13.0.2
 
 ### Patch Changes