+++ packages/select-with-tags/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @alfalab/core-components-select-with-tags
 
+## 10.1.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.1.3
+    - @alfalab/core-components-bottom-sheet@8.1.3
+
+## 10.1.2
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.1.2
+
+## 10.1.1
+
+### Patch Changes
+
+<sup><time>06.03.2026</time></sup>
+
+#### [#2102](https://github.com/core-ds/core-components/pull/2102)
+
+##### Vars
+
+- Удален файл `no-typography-index.css`
+
+- Обновлены зависимости
+    - @alfalab/core-components-select@19.1.1
+    - @alfalab/core-components-bottom-sheet@8.1.2
+    - @alfalab/core-components-form-control@14.0.4
+    - @alfalab/core-components-popover@8.0.3
+    - @alfalab/core-components-input@17.1.3
+    - @alfalab/core-components-tag@10.0.3
+    - @alfalab/core-components-mq@6.0.3
+
+## 10.1.0
+
+### Minor Changes
+
+<sup><time>27.02.2026</time></sup>
+
+#### [#2071](https://github.com/core-ds/core-components/pull/2071)
+
+##### Select
+
+- Добавлен проп `checkmarkPosition` для управления позицией чекмарка относительно контента ('before' | 'after')
+- Компонент `BaseCheckmark` теперь поддерживает проп `content` для передачи контента опции
+- Оптимизирована логика рендеринга чекбоксов и иконок в компонентах `OptionBase` и `BaseOption`
+
+##### SelectWithTags
+
+- Добавлена поддержка динамического позиционирования чекмарков в зависимости от `view` (desktop/mobile)
+- Для desktop по умолчанию используется `checkmarkPosition='before'`, для mobile - `checkmarkPosition='after'`
+- Возможность переопределить позицию через проп `optionProps.checkmarkPosition`
+
+### Patch Changes
+
+<sup><time>27.02.2026</time></sup>
+
+#### [#2053](https://github.com/core-ds/core-components/pull/2053)
+
+##### Input, AmountInput, InputAutocomplete, NumberInput, Select, FormControl, SelectWithTags
+
+- Небольшие улучшения кодовой базы компонентов
+
+- Обновлены зависимости
+    - @alfalab/core-components-input@17.1.2
+    - @alfalab/core-components-select@19.1.0
+    - @alfalab/core-components-form-control@14.0.3
+    - @alfalab/core-components-bottom-sheet@8.1.1
+
 ## 10.0.6
 
 ### Patch Changes